@@ -4,7 +4,7 @@ class Admin::BaseController < ApplicationController
   before_filter :signed_in_user, :require_admin
 
   def require_admin
-    unless current_user.is_admin?
+    unless current_user.can_admin?
       flash[:error] = "You don't have permission to do that!"
       redirect_to root_url
     end
