@@ -1,6 +1,6 @@
 require 'exception_notifier'
 
-SciRate::Application.configure do
+Rails.application.configure do
   # Settings specified here will take precedence over those in config/application.rb
 
   # Code is not reloaded between requests
@@ -13,6 +13,7 @@ SciRate::Application.configure do
   config.eager_load = true
 
   # Full error reports are disabled and caching is turned on
+  #
   config.consider_all_requests_local       = false
   config.action_controller.perform_caching = true
 
@@ -22,16 +23,19 @@ SciRate::Application.configure do
   # config.action_dispatch.rack_cache = true
 
   # Disable Rails's static asset server (Apache or nginx will already do this)
-  config.serve_static_assets = true
+  config.serve_static_files = true
 
   config.assets.js_compressor = :uglifier
   config.assets.css_compressor = :sass
 
+  # Asset digests allow you to set far-future HTTP expiration dates on all assets,
+  # yet still be able to expire them through the digest params.
+  config.assets.digest = true
+
   # Do not fallback to assets pipeline if a precompiled asset is missed
   config.assets.compile = false
 
-  # Generate digests for assets URLs
-  config.assets.digest = true
+  # `config.assets.precompile` and `config.assets.version` have moved to config/initializers/assets.rb
 
   # Version of your assets, change this if you want to expire all your assets.
   config.assets.version = '1.0'
@@ -58,9 +62,6 @@ SciRate::Application.configure do
   # Enable serving of images, stylesheets, and JavaScripts from an asset server
   # config.action_controller.asset_host = "http://assets.example.com"
 
-  # Precompile additional assets (application.js, application.css, and all non-JS/CSS are already added)
-  config.assets.precompile << /\.(?:svg|eot|woff|ttf)$/
-
   # Ignore bad email addresses and do not raise email delivery errors.
   # Set this to true and configure the email server for immediate delivery to raise delivery errors.
   # config.action_mailer.raise_delivery_errors = false
@@ -72,27 +73,11 @@ SciRate::Application.configure do
   # Send deprecation notices to registered listeners
   config.active_support.deprecation = :notify
 
-  # Disable automatic flushing of the log to improve performance.
-  # config.autoflush_log = false
-
   # Use default logging formatter so that PID and timestamp are not suppressed.
   config.log_formatter = ::Logger::Formatter.new
 
   # Do not dump schema after migrations.
   config.active_record.dump_schema_after_migration = false
-
-  ActionMailer::Base.delivery_method = :smtp
-  ActionMailer::Base.perform_deliveries = true
-  ActionMailer::Base.raise_delivery_errors = true
-  ActionMailer::Base.smtp_settings = {
-    :address => "smtp.gmail.com",
-    :port => "587",
-    :domain => "gmail.com",
-    :enable_starttls_auto => true,
-    :authentication => :login,
-    :user_name => Settings::GMAIL_SMTP_USER,
-    :password => Settings::GMAIL_SMTP_PASSWORD
-  }
 
   WillPaginate.per_page = 100
 
