@@ -1,7 +1,7 @@
 require 'sessions_helper'
 
 SciRate::Application.routes.draw do
-  get '/', to: 'feeds#index', as: 'root'
+  root 'feeds#index'
 
   get '/search', to: 'papers#search', as: 'papers_search'
 
@@ -70,10 +70,9 @@ SciRate::Application.routes.draw do
   get '/users/:id/activate/:confirmation_token', to: 'users#activate', as: 'activate_user'
   get '/feeds', to: 'users#feeds', as: 'feeds'
 
-  get '/arxiv/:paper_uid/scites', to: 'papers#scites', paper_uid: /.+\/.+|\d+.\d+(v\d)?/, as: 'paper_scites'
+  get '/arxiv/:paper_uid/scites', to: 'papers#scites', paper_uid: /(.+\/.+|\d+.\d+)(v\d)?/, as: 'paper_scites'
   get '/arxiv/:feed/comments', to: 'comments#index', feed: /.+/, as: 'feed_comments'
-  get '/arxiv/:paper_uid', to: 'papers#show', paper_uid: /.+\/.+|\d+.\d+(v\d)?/, as: 'paper'
-
+  get '/arxiv/*paper_uid', to: 'papers#show', paper_uid: /.+\/.+|\d+.\d+(v\d)?/, as: 'paper'
   get '/arxiv/:feed', to: 'feeds#show', feed: /.+/, as: 'feed'#, constraints: NeedsUserConstraint
 
   get '/admin', to: 'admin/base#index', as: 'admin'
@@ -87,62 +86,4 @@ SciRate::Application.routes.draw do
   get '/:username/comments', to: 'users#comments', username: /.+/, as: 'user_comments'
   get '/:username/papers', to: 'users#papers', username: /.+/, as: 'user_papers'
   get '/:username', to: 'users#activity', username: /.+/, as: 'user'
-
-
-  # The priority is based upon order of creation:
-  # first created -> highest priority.
-
-  # Sample of regular route:
-  #   get 'products/:id' => 'catalog#view'
-  # Keep in mind you can assign values other than :controller and :action
-
-  # Sample of named route:
-  #   get 'products/:id/purchase' => 'catalog#purchase', :as => :purchase
-  # This route can be invoked with purchase_url(:id => product.id)
-
-  # Sample resource route (maps HTTP verbs to controller actions automatically):
-  #   resources :products
-
-  # Sample resource route with options:
-  #   resources :products do
-  #     member do
-  #       get 'short'
-  #       post 'toggle'
-  #     end
-  #
-  #     collection do
-  #       get 'sold'
-  #     end
-  #   end
-
-  # Sample resource route with sub-resources:
-  #   resources :products do
-  #     resources :comments, :sales
-  #     resource :seller
-  #   end
-
-  # Sample resource route with more complex sub-resources
-  #   resources :products do
-  #     resources :comments
-  #     resources :sales do
-  #       get 'recent', :on => :collection
-  #     end
-  #   end
-
-  # Sample resource route within a namespace:
-  #   namespace :admin do
-  #     # Directs /admin/products/* to Admin::ProductsController
-  #     # (app/controllers/admin/products_controller.rb)
-  #     resources :products
-  #   end
-
-  # You can have the root of your site routed with "root"
-  # just remember to delete public/index.html.
-  # root :to => 'welcome#index'
-
-  # See how all your routes lay out with "rake routes"
-
-  # This is a legacy wild controller route that's not recommended for RESTful applications.
-  # Note: This route will make all actions in every controller accessible via GET requests.
-  # get ':controller(/:action(/:id))(.:format)'
 end
