@@ -12,7 +12,7 @@
 require 'spec_helper'
 
 describe CommentReport do
-  it { should belong_to(:user) }
-  it { should belong_to(:comment) }
-  it { should validate_uniqueness_of(:comment_id).scoped_to(:user_id) }
+  it { is_expected.to belong_to(:user) }
+  it { is_expected.to belong_to(:comment) }
+  it { is_expected.to validate_uniqueness_of(:comment_id).scoped_to(:user_id) }
 end
