@@ -34,31 +34,36 @@ describe Paper do
 
   subject { paper }
 
-  it { should respond_to(:title) }
-  it { should respond_to(:authors) }
-  it { should respond_to(:abstract) }
-  it { should respond_to(:uid) }
-  it { should respond_to(:abs_url) }
-  it { should respond_to(:pdf_url) }
-  it { should respond_to(:submit_date) }
-  it { should respond_to(:update_date) }
-  it { should respond_to(:scites) }
-  it { should respond_to(:sciters) }
-  it { should respond_to(:comments) }
-  it { should respond_to(:categories) }
+  it { is_expected.to respond_to(:title) }
+  it { is_expected.to respond_to(:authors) }
+  it { is_expected.to respond_to(:abstract) }
+  it { is_expected.to respond_to(:uid) }
+  it { is_expected.to respond_to(:abs_url) }
+  it { is_expected.to respond_to(:pdf_url) }
+  it { is_expected.to respond_to(:submit_date) }
+  it { is_expected.to respond_to(:update_date) }
+  it { is_expected.to respond_to(:scites) }
+  it { is_expected.to respond_to(:sciters) }
+  it { is_expected.to respond_to(:comments) }
+  it { is_expected.to respond_to(:categories) }
 
-  it { should validate_uniqueness_of(:uid) }
-  it { should be_valid }
+  it { is_expected.to validate_uniqueness_of(:uid) }
+  it { is_expected.to be_valid }
 
   describe "#to_bibtex" do
     let(:paper) do
-      FactoryGirl.create(:paper_with_versions,
+      paper = FactoryGirl.create(:paper_with_versions,
         author_str: "Susanta K. Khan and Madhumangal Pal",
         title: "Interval-Valued Intuitionistic Fuzzy Matrices",
         pubdate: "2014-05-12 01:00:00 UTC",
         uid: "1404.6949",
         journal_ref: "Notes on Intuitionistic Fuzzy Sets, 11(1) (2005)16-27"
       )
+
+      FactoryGirl.create(:author, fullname: "Susanta K. Khan", paper_uid: paper.uid)
+      FactoryGirl.create(:author, fullname: "Madhumangal Pal", paper_uid: paper.uid)
+
+      paper
     end
 
     it "generates bibtex correctly" do
