@@ -34,7 +34,17 @@ Dir[Rails.root.join("spec/support/**/*.rb")].each {|f| require f}
 # If you are not using ActiveRecord, you can remove this line.
 ActiveRecord::Migration.check_pending! if defined?(ActiveRecord::Migration)
 
-Capybara.javascript_driver = :webkit
+Capybara.configure do |config|
+  Capybara.register_driver :webkit do |app|
+    driver = Capybara::Webkit::Driver.new(app)
+    driver.allow_url("cdn.mathjax.org")
+    driver.allow_url("www.google-analytics.com")
+    driver
+  end
+
+  config.javascript_driver = :webkit
+end
+
 
 RSpec.configure do |config|
   config.infer_spec_type_from_file_location!
